Move MutexGuardedObject to its own file.
